# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Legal Case Context Builder API - Build context for legal cases using directed acyclic graphs\n\nThis microservice provides a context builder for legal cases using directed acyclic graphs (DAGs).\nUsers traverse through question nodes to build comprehensive case context with evidence tracking,\ntimeline management, and legal assessment capabilities.",
        "title": "Jurigen API",
        "contact": {
            "name": "Legal Tech Support",
            "email": "support@legaltech.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/dags": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all available Legal Case DAG identifiers in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAGs"
                ],
                "summary": "List Legal Case DAGs",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved DAG list",
                        "schema": {
                            "$ref": "#/definitions/http.DAGListPresenter"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dags/validate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Validate a DAG structure to ensure it meets all requirements (single root node, acyclic, valid relationships)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAGs"
                ],
                "summary": "Validate Legal Case DAG",
                "parameters": [
                    {
                        "description": "DAG structure to validate",
                        "name": "dag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DAG validation completed (may contain errors)",
                        "schema": {
                            "$ref": "#/definitions/http.ValidationResultPresenter"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dags/{dagId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a complete Legal Case DAG structure including all questions, answers, and collected context",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAGs"
                ],
                "summary": "Get Legal Case DAG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG unique identifier (UUID)",
                        "name": "dagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved DAG",
                        "schema": {
                            "$ref": "#/definitions/http.DAGPresenter"
                        }
                    },
                    "400": {
                        "description": "Invalid DAG ID format",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "DAG not found",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a complete Legal Case DAG structure including questions, answers, and context",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAGs"
                ],
                "summary": "Update Legal Case DAG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG unique identifier (UUID)",
                        "name": "dagId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated DAG structure",
                        "name": "dag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.DAGPresenter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated DAG",
                        "schema": {
                            "$ref": "#/definitions/http.DAGPresenter"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or DAG ID format",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "DAG not found",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xhttp.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.AnswerPresenter": {
            "description": "An answer to a legal question with optional user context and structured metadata for evidence tracking",
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "Yes, age discrimination occurred"
                },
                "id": {
                    "type": "string",
                    "example": "fc28c4b6-d185-cf56-a7e4-dead499ff1e8"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "next_node": {
                    "type": "string",
                    "example": "8b007ce4-b676-5fb3-9f93-f5f6c41cb655"
                },
                "user_context": {
                    "type": "string",
                    "example": "Manager explicitly mentioned my age during termination"
                }
            }
        },
        "http.DAGListPresenter": {
            "description": "List of Legal Case DAG identifiers available in the system",
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "dags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "http.DAGPresenter": {
            "description": "Legal Case DAG with questions, answers, and context",
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.NodePresenter"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Employment Discrimination Case"
                }
            }
        },
        "http.NodePresenter": {
            "description": "A question node with potential answers for legal case context building",
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.AnswerPresenter"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "8b007ce4-b676-5fb3-9f93-f5f6c41cb655"
                },
                "question": {
                    "type": "string",
                    "example": "Were you discriminated against in the workplace?"
                }
            }
        },
        "http.ValidateRequest": {
            "description": "DAG validation request containing the DAG structure to validate",
            "type": "object",
            "required": [
                "dag"
            ],
            "properties": {
                "dag": {
                    "$ref": "#/definitions/http.DAGPresenter"
                }
            }
        },
        "http.ValidationErrorPresenter": {
            "description": "Detailed validation error with context information",
            "type": "object",
            "properties": {
                "answer_id": {
                    "type": "string",
                    "example": "fc28c4b6-d185-cf56-a7e4-dead499ff1e8"
                },
                "code": {
                    "type": "string",
                    "example": "DAG_HAS_CYCLES"
                },
                "message": {
                    "type": "string",
                    "example": "DAG contains 1 cycle(s). A valid DAG must be acyclic"
                },
                "node_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "severity": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "http.ValidationResultPresenter": {
            "description": "Comprehensive DAG validation results including errors, warnings, and statistics",
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ValidationErrorPresenter"
                    }
                },
                "is_valid": {
                    "type": "boolean",
                    "example": true
                },
                "statistics": {
                    "$ref": "#/definitions/http.ValidationStatisticsPresenter"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ValidationWarningPresenter"
                    }
                }
            }
        },
        "http.ValidationStatisticsPresenter": {
            "description": "Statistical information about the DAG structure and validation results",
            "type": "object",
            "properties": {
                "cycle_paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "has_cycles": {
                    "type": "boolean",
                    "example": false
                },
                "leaf_node_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "leaf_nodes": {
                    "type": "integer",
                    "example": 2
                },
                "max_depth": {
                    "type": "integer",
                    "example": 3
                },
                "root_node_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "root_nodes": {
                    "type": "integer",
                    "example": 1
                },
                "total_answers": {
                    "type": "integer",
                    "example": 12
                },
                "total_nodes": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "http.ValidationWarningPresenter": {
            "description": "Non-critical validation issue that doesn't prevent DAG usage",
            "type": "object",
            "properties": {
                "answer_id": {
                    "type": "string",
                    "example": "fc28c4b6-d185-cf56-a7e4-dead499ff1e8"
                },
                "code": {
                    "type": "string",
                    "example": "DEEP_NESTING"
                },
                "message": {
                    "type": "string",
                    "example": "DAG has deep nesting which may impact performance"
                },
                "node_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "xhttp.ErrorResponse": {
            "description": "Standard error response format for API failures",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "DAG not found"
                },
                "message": {
                    "type": "string",
                    "example": "failed to get DAG"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "tags": [
        {
            "description": "Operations for managing and retrieving Legal Case DAGs",
            "name": "DAGs"
        }
    ]
}